+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 2.0 - 2018-06-28
+
+### Added
+
+- A new data source to de-serialize a list of strings - `yaml_list_of_strings`
+
+### Changed
+
+- `yaml` renamed to `yaml_map_of_strings`
+
+
 ## 1.0 - 2018-06-20
 
 ### Added